Moving GDPR specifics into GDPR namespace
