+++ Source/Fundamentals/Compliance/GDPR/IHoldPII.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Compliance
+namespace Cratis.Compliance.GDPR
 {
     /// <summary>
     /// Defines a marker interface for types to indicate the type holds PII defined by GDPR.
