+++ Source/Fundamentals/Compliance/GDPR/PIIAttribute.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Compliance
+namespace Cratis.Compliance.GDPR
 {
     /// <summary>
     /// Represents an attribute that can be used to mark classes or properties to indicate the information kept is PII according to the definition of GDPR.
