+++ Specifications/Fundamentals/Compliance/GDPR/for_PIIMetadataProvider/given/a_provider.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Compliance.for_PIIMetadataProvider.given
+namespace Cratis.Compliance.GDPR.for_PIIMetadataProvider.given
 {
     public class a_provider : Specification
     {
