+++ Specifications/Fundamentals/Compliance/GDPR/for_PIIMetadataProvider/when_asking_if_can_provide_for_property/and_declaring_type_is_adorned_with_pii_attribute.cs
@@ -3,13 +3,13 @@
 
 using System.Reflection;
 
-namespace Cratis.Compliance.for_PIIMetadataProvider.when_asking_if_can_provide_for_property
+namespace Cratis.Compliance.GDPR.for_PIIMetadataProvider.when_asking_if_can_provide_for_property
 {
-    public class and_property_is_adorned_with_pii_attribute : given.a_provider
+    public class and_declaring_type_is_adorned_with_pii_attribute : given.a_provider
     {
+        [PII]
         class MyClass
         {
-            [PII]
             public string Something { get; set; }
 
             public static PropertyInfo SomethingProperty = typeof(MyClass).GetProperty(nameof(Something), BindingFlags.Public | BindingFlags.Instance);
