+++ Specifications/Fundamentals/Compliance/GDPR/for_PIIMetadataProvider/when_asking_if_can_provide_for_property/and_property_is_adorned_with_pii_attribute.cs
@@ -3,20 +3,21 @@
 
 using System.Reflection;
 
-namespace Cratis.Compliance.for_PIIMetadataProvider.when_providing_for_property
+namespace Cratis.Compliance.GDPR.for_PIIMetadataProvider.when_asking_if_can_provide_for_property
 {
-    public class and_there_is_no_metadata : given.a_provider
+    public class and_property_is_adorned_with_pii_attribute : given.a_provider
     {
         class MyClass
         {
+            [PII]
             public string Something { get; set; }
 
             public static PropertyInfo SomethingProperty = typeof(MyClass).GetProperty(nameof(Something), BindingFlags.Public | BindingFlags.Instance);
         }
 
-        Exception result;
-        void Because() => result = Catch.Exception(() => provider.Provide(MyClass.SomethingProperty));
+        bool result;
+        void Because() => result = provider.CanProvide(MyClass.SomethingProperty);
 
-        [Fact] void should_throw_no_compliance_metadata_for_property() => result.ShouldBeOfExactType<NoComplianceMetadataForProperty>();
+        [Fact] void should_be_able_to_provide() => result.ShouldBeTrue();
     }
 }
