+++ Specifications/Fundamentals/Compliance/GDPR/for_PIIMetadataProvider/when_providing_for_type/and_there_is_metadata_and_details.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Compliance.for_PIIMetadataProvider.when_providing_for_type
+namespace Cratis.Compliance.GDPR.for_PIIMetadataProvider.when_providing_for_type
 {
     public class and_there_is_metadata_and_details : given.a_provider
     {
